refactored graph into component
